Fix small bug; update comments re: testability
Also add test_mblog.py.

Also note issues with getrevid, gettimestamp; should be handled better
in the future, but for now it will raise an error (if this happens,
matchbot tried to post a normal edit to a Flow board; ought to be fixed
but should still be an error).